requirements-light.txt
llama-cpp-python>=0.2.79
huggingface-hub>=0.21.0
numpy>=1.24.0
requests>=2.31.0
gguf>=0.6.0
python-dotenv>=1.0.0
